Scoring service

+++ src/main/java/com/vins/spring_boot_training/service/FibonacciService.java
@@ -6,17 +6,17 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class FibonacciService {
-  public int getFibonacci(int n) {
+  public Long getFibonacci(long n) {
     if (n < 0) {
       throw new CustomException(FibonacciErrors.NEGATIVE_NUMBER);
     }
 
     if (n <= 1) return n;
 
-    int a = 0;
-    int b = 1;
-    for (int i = 2; i <= n; i++) {
-      int next = a + b;
+    long a = 0;
+    long b = 1;
+    for (long i = 2; i <= n; i++) {
+      long next = a + b;
       a = b;
       b = next;
     }
